輕微調整
