+++ src/pkg/utils/utils.ts
@@ -236,3 +236,21 @@ export function strToBase64(str: string): string {
   );
 }
 */
+
+export function getMetadataStr(code: string): string | null {
+  const start = code.indexOf("==UserScript==");
+  const end = code.indexOf("==/UserScript==");
+  if (start === -1 || end === -1) {
+    return null;
+  }
+  return `// ${code.substring(start, end + 15)}`;
+}
+
+export function getUserConfigStr(code: string): string | null {
+  const start = code.indexOf("==UserConfig==");
+  const end = code.indexOf("==/UserConfig==");
+  if (start === -1 || end === -1) {
+    return null;
+  }
+  return `/* ${code.substring(start, end + 15)} */`;
+}
